do not persist users and groups to chef node
The list of users and groups that were being kept were automatically
being persisted to the chef server via the node object. This removes
those attributes to ensure that they're not present in the node object.

The list does not need to be persisted as it is generated statically on
each run.